apps/api/app/core/security_headers.py
from starlette.middleware.base import BaseHTTPMiddleware
from fastapi import Request


class SecurityHeadersMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        response = await call_next(request)

        response.headers["X-Content-Type-Options"] = "nosniff"
        response.headers["X-Frame-Options"] = "DENY"
        response.headers["X-XSS-Protection"] = "1; mode=block"
        response.headers["Referrer-Policy"] = "no-referrer"

        # Solo activar HSTS en producción
        # (HTTPS obligatorio)
        if request.url.scheme == "https":
            response.headers["Strict-Transport-Security"] = \
                "max-age=63072000; includeSubDomains"

        return response
