apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

pydantic==2.6.4
pydantic-settings==2.2.1

motor==3.3.2
pymongo==4.6.2

redis==5.0.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

jinja2==3.1.3
aiohttp==3.9.3
aiosmtplib==2.0.2
email-validator==2.1.1

jinja2==3.1.3